+++ sample/iam/example_iam_create_apikey_permanently_valid.go
@@ -1,6 +1,7 @@
 package iamsample
 
 import (
+	"encoding/json"
 	"fmt"
 	"github.com/baidubce/baiducloud-go-sdk/core/util"
 	"github.com/baidubce/baiducloud-go-sdk/services/iam"
@@ -14,14 +15,26 @@ func CreateApikeyPermanentlyValid() {
 		fmt.Println("create client err:", err)
 		return
 	}
+	Acl := &iam.ACL{
+		Id:                util.PtrString(""),
+		Version:           util.PtrString(""),
+		AccessControlList: []*iam.ACLEntry{},
+	}
 	createApikeyPermanentlyValidRequest := &iam.CreateApikeyPermanentlyValidRequest{
 		UserId: util.PtrString(""),
-		Acl:    util.PtrString(""),
+		Acl:    Acl,
 		Name:   util.PtrString(""),
 	}
-	err = client.CreateApikeyPermanentlyValid(createApikeyPermanentlyValidRequest)
+	result, err := client.CreateApikeyPermanentlyValid(createApikeyPermanentlyValidRequest)
 	if err != nil {
 		// 此处仅做打印展示，请谨慎对待异常处理，在工程项目中切勿直接忽略异常。
 		fmt.Println("request failed:", err)
+		return
 	}
+	data, err := json.MarshalIndent(result, "", "    ")
+	if err != nil {
+		fmt.Println("json marshalIndent failed:", err)
+		return
+	}
+	fmt.Println(string(data))
 }