+++ sample/iam/example_iam_delete_apikey_permanently_valid.go
@@ -1,6 +1,7 @@
 package iamsample
 
 import (
+	"encoding/json"
 	"fmt"
 	"github.com/baidubce/baiducloud-go-sdk/core/util"
 	"github.com/baidubce/baiducloud-go-sdk/services/iam"
@@ -18,9 +19,16 @@ func DeleteApikeyPermanentlyValid() {
 		UserId: util.PtrString(""),
 		Id:     util.PtrString(""),
 	}
-	err = client.DeleteApikeyPermanentlyValid(deleteApikeyPermanentlyValidRequest)
+	result, err := client.DeleteApikeyPermanentlyValid(deleteApikeyPermanentlyValidRequest)
 	if err != nil {
 		// 此处仅做打印展示，请谨慎对待异常处理，在工程项目中切勿直接忽略异常。
 		fmt.Println("request failed:", err)
+		return
 	}
+	data, err := json.MarshalIndent(result, "", "    ")
+	if err != nil {
+		fmt.Println("json marshalIndent failed:", err)
+		return
+	}
+	fmt.Println(string(data))
 }