+++ services/iam/model_create_apikey_permanently_valid_request.go
@@ -2,6 +2,6 @@ package iam
 
 type CreateApikeyPermanentlyValidRequest struct {
 	UserId *string `json:"userId,omitempty"`
-	Acl    *string `json:"acl,omitempty"`
+	Acl    *ACL    `json:"acl,omitempty"`
 	Name   *string `json:"name,omitempty"`
 }