+++ services/iam/model_delete_apikey_permanently_valid_response.go
@@ -1,0 +1,9 @@
+package iam
+
+import "github.com/baidubce/baiducloud-go-sdk/bce"
+
+type DeleteApikeyPermanentlyValidResponse struct {
+	bce.BaseResponse
+	Success *bool  `json:"success,omitempty"`
+	Status  *int32 `json:"status,omitempty"`
+}