+++ services/iam/model_obtain_a_list_of_permanently_valid_apikeys_response.go
@@ -1,0 +1,10 @@
+package iam
+
+import "github.com/baidubce/baiducloud-go-sdk/bce"
+
+type ObtainAListOfPermanentlyValidApikeysResponse struct {
+	bce.BaseResponse
+	Success *bool  `json:"success,omitempty"`
+	Status  *int32 `json:"status,omitempty"`
+	Page    *Page  `json:"page,omitempty"`
+}