+++ services/iam/model_update_apikey_permanently_valid_request.go
@@ -3,5 +3,5 @@ package iam
 type UpdateApikeyPermanentlyValidRequest struct {
 	UserId *string `json:"userId,omitempty"`
 	Id     *string `json:"id,omitempty"`
-	Acl    *string `json:"acl,omitempty"`
+	Acl    *ACL    `json:"acl,omitempty"`
 }